backend/pyproject.toml
[project]
name = "coco-ai-backend"
version = "0.1.0"
description = "Backend for Coco-Ai using Google ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # エージェントグラフを構築するためのコアフレームワーク
    "google-adk>=1.12.0",
 
    # Webフレームワーク
    "fastapi",

    # EventarcからのCloudEventsを処理するために必要
    "cloudevents",
 
    # 型安全な設定管理のために必要
    "pydantic-settings",
 
    # agents/services で使用されるGoogle Cloudサービスの直接的な依存関係
    "google-cloud-firestore",
    "google-cloud-speech",
    "google-cloud-texttospeech",
    "google-cloud-storage",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    # リンターとフォーマッターを兼ねる
    "ruff",

    # ASGIサーバー (FastAPIのローカル実行に必要)
    "uvicorn[standard]",
]
